+++ tap_intercom/streams.py
@@ -16,6 +16,15 @@
 #   interpolate_page: True, False (to determine start page based on total pages and bookmark)
 #       default: False
 
+def build_query(body, value, starting_after=None):
+    body['query']['value'] = value
+    if starting_after:
+        body['pagination'] = {
+            'starting_after': starting_after
+        }
+    return body
+
+
 STREAMS = {
     'admin_list': {
         'path': 'admins',
@@ -94,7 +103,19 @@ STREAMS = {
         }
     },
     'contacts': {
-        'path': 'contacts',
+        'path': 'contacts/search',
+        'method': 'POST',
+        'search_query': {
+            'query': {
+                'field': 'updated_at',
+                'operator': '>',
+                'value': 'value'
+            },
+            'sort': {
+                'field': 'updated_at',
+                'order': 'ascending'
+            }
+        },
         'data_key': 'data',
         'key_properties': ['id'],
         'replication_method': 'INCREMENTAL',
@@ -102,7 +123,7 @@ STREAMS = {
         'bookmark_type': 'datetime',
         'batch_size': 150,
         # V2 APIs are starting to adopt a cursor-based approach to pagination
-        'cursor': True
+        'search': True
     },
     'segments': {
         'key_properties': ['id'],