+++ tap_intercom/transform.py
@@ -1,3 +1,8 @@
+import math as m
+
+from singer.utils import strptime_to_utc
+
+
 # De-nest each list node up to record level
 def denest_list_nodes(this_json, data_key, list_nodes):
     new_json = this_json
@@ -52,3 +57,32 @@ def transform_json(this_json, stream_name, data_key):
     if data_key in new_json:
         return new_json[data_key]
     return new_json
+
+
+def find_datetimes_in_schema(schema):
+    datetimes = []
+    for element, value in schema['properties'].items():
+        if 'format' in value and value['format'] == 'date-time':
+            datetimes.append(element)
+    return datetimes
+
+def get_integer_places(value):
+    if value <= 999999999999997:
+        return int(m.log10(value)) + 1
+    counter = 15
+    while value >= 10**counter:
+        counter += 1
+    return counter
+
+
+# API returns date times, epoch seconds and epoch millis
+# Transform datetimes to epoch millis
+# Transform epoch seconds to millis
+def transform_times(record, schema_datetimes):
+    for datetime in schema_datetimes:
+        if datetime in record and record[datetime]:
+            if isinstance(record[datetime], str):
+                record[datetime] = strptime_to_utc(
+                    record[datetime]).timestamp() * 1000
+            elif get_integer_places(record[datetime]) == 10:
+                record[datetime] = record[datetime] * 1000