+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.2
+  * Support date time strings from API in tranform
+
+## 1.0.1
+  * Fix for inconsistem epoch formats from API. Normalize all to milli.
+
+## 1.0.0
+  * Bumping to first major version 1.0.0 for GA Release
+
 ## 0.1.0
   * Add `sla_applied` and `statistics` objects.